[CSL-1122] Fix Tests [WIP]

+++ test/Test/Pos/Txp/Toil/UtxoSpec.hs
@@ -62,7 +62,7 @@ spec = describe "Txp.Toil.Utxo" $ do
     scriptTxSpec
   where
     myTx = TxIn myHash 0
-    myHash = unsafeHash @Int64 0
+    myHash = unsafeHash @Int32 0
     description_findTxInUtxo =
         "correctly finds the TxOut corresponding to (txHash, txIndex) when the key is in\
         \ the Utxo map, and doesn't find it otherwise"
